updated based on pr comments from jeff
