+++ packages/wonder-stuff-ci/src/index.ts
@@ -1,6 +1,6 @@
-export {
-    getMobileReleaseTags,
-    extractMobileReleaseBranch,
-    compareVersions,
-} from "./mobile-release-utils";
-export {execProm, bufferToString} from "./exec-prom";
+export {compareVersions} from "./compare-versions";
+export {extractMobileReleaseInfoFromBranchName} from "./extract-mobile-release-info-from-branch-name";
+export {getMobileReleaseTags} from "./get-mobile-release-tags";
+export {getTagsFromGit} from "./get-tags-from-git";
+export {execAsync} from "./exec-async";
+export {bufferToString} from "./buffer-to-string";
